Add query planning strategy objects
