+++ lib/sycamore/sycamore/query/planner.py
@@ -1,43 +1,26 @@
-from dataclasses import dataclass
 import logging
 import typing
+from dataclasses import dataclass
 from typing import Any, List, Optional, Tuple, Type
 
 from sycamore.llms.llms import LLM
 from sycamore.llms.openai import OpenAI, OpenAIModels
 from sycamore.query.logical_plan import LogicalPlan, Node
-from sycamore.query.operators.count import Count
-from sycamore.query.operators.llm_filter import LlmFilter
 from sycamore.query.operators.basic_filter import BasicFilter
-from sycamore.query.operators.llm_extract_entity import LlmExtractEntity
-from sycamore.query.operators.query_database import QueryDatabase, QueryVectorDatabase
-from sycamore.query.operators.math import Math
-from sycamore.query.operators.sort import Sort
-from sycamore.query.operators.summarize_data import SummarizeData
-from sycamore.query.operators.top_k import TopK
+from sycamore.query.operators.count import Count
 from sycamore.query.operators.limit import Limit
+from sycamore.query.operators.llm_extract_entity import LlmExtractEntity
+from sycamore.query.operators.llm_filter import LlmFilter
+from sycamore.query.operators.math import Math
+from sycamore.query.operators.query_database import QueryDatabase, QueryVectorDatabase
+from sycamore.query.operators.sort import Sort
+from sycamore.query.operators.top_k import TopK
 from sycamore.query.schema import OpenSearchSchema, OpenSearchSchemaField
+from sycamore.query.strategy import QueryPlanStrategy, ALL_OPERATORS
 from sycamore.utils.extract_json import extract_json
 
 if typing.TYPE_CHECKING:
     from opensearchpy import OpenSearch
-
-
-# All operators that are allowed for construction of a query plan.
-# If a class is not in this list, it will not be used.
-OPERATORS: List[Type[Node]] = [
-    QueryDatabase,
-    QueryVectorDatabase,
-    BasicFilter,
-    LlmFilter,
-    LlmExtractEntity,
-    Count,
-    SummarizeData,
-    Math,
-    Sort,
-    TopK,
-    Limit,
-]
 
 
 # This is the base prompt for the planner.
@@ -56,11 +39,10 @@ guidelines when generating a plan:
             "properties.key" or "properties.count"; you can only reference one of those fields.
             Other than those, DO NOT USE ANY OTHER FIELD NAMES.
         5. If an optional field does not have a value in the query plan, return null in its place.
-        6. The first step of each plan MUST be a **QueryDatabase** or **QueryVectorDatabase" operation that returns a 
-           database. Whenever possible, include all possible filtering operations in the QueryDatabase step.
+        6. The first step of each plan MUST be a **QueryDatabase** or **QueryVectorDatabase" operation. 
+            Whenever possible, include all possible filtering operations in the first step.
            That is, you should strive to construct an OpenSearch query that filters the data as
-           much as possible, reducing the need for further query operations. Use a QueryVectorDatabase step instead of
-           an LLMFilter if the query looks reasonable.
+           much as possible, reducing the need for further query operations.
 """
 
 # Variants on the last step in the query plan, based on whether the user has requested raw data
@@ -78,104 +60,6 @@ def process_json_plan(json_plan: str) -> LogicalPlan:
     if not isinstance(parsed_plan, dict):
         raise ValueError(f"Expected LLM query plan to contain a dict, got f{type(parsed_plan)}")
     return LogicalPlan.model_validate(parsed_plan)
-
-
-def postprocess_llm_helper(user_message: str, llm_client: LLM) -> str:
-    messages = [
-        {
-            "role": "system",
-            "content": """You are a helpful agent that assists in small transformations
-                    of input text as per the instructions. You should make minimal changes
-                    to the provided input and keep your response short""",
-        },
-        {"role": "user", "content": user_message},
-    ]
-
-    prompt_kwargs = {"messages": messages}
-    chat_completion = llm_client.generate(prompt_kwargs=prompt_kwargs, llm_kwargs={"temperature": 0, "seed": 42})
-    return chat_completion
-
-
-def postprocess_plan(plan: Any, llm_client: LLM) -> Any:
-    """Given the LogicalPlan query plan provided by the LLM, postprocess it using a set of rules
-    that modify the plan for optimizing or other purposes."""
-
-    # Rule 1: If the plan has a vector search in the beginning followed by a count or nothing or extract_entity,
-    # we replace that by removing the vector search and adding an LLM Filter before the count
-
-    if plan.nodes[0].node_type == "QueryVectorDatabase" and (
-        len(plan.nodes) == 1 or plan.nodes[1].node_type in ["Count", "LlmExtractEntity", "SummarizeData"]
-    ):
-        # If the first operator has an "opensearch_filter", we will convert it to a QueryDatabase with
-        #        that opensearch_filter as "query"
-        # If not, we do a QueryDatabase with "match_all" instead
-        op = plan.nodes[0]
-
-        modified_description = postprocess_llm_helper(
-            f"""
-                    The following is the description of a Python function. I am modifying the function code
-                    to remove any functionality that specifically has to do with "{op.query_phrase}", thereby 
-                    generalizing the description to be more flexible.
-                    Return only the modified description. Do not make assumptions
-                    about the intent of the question that are not explicitly specified.
-                    {op.description}
-                    """,
-            llm_client,
-        )
-
-        new_op = QueryDatabase.model_validate(
-            {
-                "node_id": op.node_id,
-                "node_type": "QueryDatabase",
-                "description": modified_description,
-                "index": op.index,
-            }
-        )
-
-        if op.opensearch_filter:
-            new_op.query = op.opensearch_filter
-        else:
-            new_op.query = {"match_all": {}}
-
-        plan.replace_node(0, new_op)
-
-        # Add an LLM Filter as the second operator
-        llm_op_description = postprocess_llm_helper(
-            f"""
-                Generate a one-line description for a python function whose goal is to filter the input
-                records based on whether they contain {op.query_phrase}.
-                Here are two example outputs: 
-                (1) Filter to records involving wildfires.
-                (2) Filter to records that occured in Northwest USA.
-                """,
-            llm_client,
-        )
-        llm_op_question = postprocess_llm_helper(
-            f"""
-                Generate a one-line true/false question that is appropriate to check whether an input document
-                satisfies {op.query_phrase}. Keep it as generic and short as possible. Do not make assumptions
-                about the intent of the question that are not explicitly specified.
-
-                Here are two examples:
-                (1) Was this incident caused by an environmental condition?
-                (2) Did this incident occur in Georgia?
-                """,
-            llm_client,
-        )
-        llm_op = LlmFilter.model_validate(
-            {
-                "node_id": 1,
-                "node_type": "LlmFilter",
-                "description": llm_op_description,
-                "inputs": [0],
-                "field": "text_representation",
-                "question": llm_op_question,
-            }
-        )
-
-        plan.insert_node(1, llm_op)
-
-    return plan
 
 
 @dataclass
@@ -461,7 +345,7 @@ class LlmPlanner:
         data_schema: A dictionary mapping field names to their types.
         os_config: The OpenSearch configuration.
         os_client: The OpenSearch client.
-        operators: A list of operators to use in the query plan.
+        strategy: Strategy to use for planning, can be used to balance cost vs speed.
         llm_client: The LLM client.
         examples: Query examples to assist the LLM planner in few-shot learning.
             You may override this to customize the few-shot examples provided to the planner.
@@ -475,7 +359,7 @@ class LlmPlanner:
         data_schema: OpenSearchSchema,
         os_config: dict[str, str],
         os_client: "OpenSearch",
-        operators: Optional[List[Type[Node]]] = None,
+        strategy: QueryPlanStrategy = QueryPlanStrategy(ALL_OPERATORS, []),
         llm_client: Optional[LLM] = None,
         examples: Optional[List[PlannerExample]] = None,
         natural_language_response: bool = False,
@@ -483,7 +367,7 @@ class LlmPlanner:
         super().__init__()
         self._index = index
         self._data_schema = data_schema
-        self._operators = operators if operators else OPERATORS
+        self._strategy = strategy
         self._os_config = os_config
         self._os_client = os_client
         self._llm_client = llm_client or OpenAI(OpenAIModels.GPT_4O.value)
@@ -558,7 +442,7 @@ class LlmPlanner:
 
         OPERATORS:
         """
-        for operator in self._operators:
+        for operator in self._strategy.operators:
             prompt += self.make_operator_prompt(operator)
 
         # Data schema.
@@ -605,7 +489,8 @@ class LlmPlanner:
         llm_prompt, llm_plan = self.generate_from_llm(question)
         try:
             plan = process_json_plan(llm_plan)
-            plan = postprocess_plan(plan, self._llm_client)
+            for processor in self._strategy.post_processors:
+                plan = processor(plan)
         except Exception as e:
             logging.error(f"Error processing LLM-generated query plan: {e}\nPlan is:\n{llm_plan}")
             raise
